pinctr: renesas: pinctrl-rzg2l: fix support for multiple gpiochips
There is a warning when booting to Linux:
"gpio gpiochip0: (11030000.pinctrl): detected irqchip that is shared
with multiple gpiochips: please fix the driver."

Refer commit "gpio: dln2: Use irqchip template" to fix this warning.
<ecb55df8f3a8f30cb188a8bd3c9d0b977eb92e09>

This makes the driver use the irqchip template to assign
properties to the gpio_irq_chip instead of using the
explicit calls to gpiochip_irqchip_add(). The irqchip is
instead added while adding the gpiochip.

Signed-off-by: hienhuynh <hien.huynh.px@renesas.com>