media: rzg2l-cru: set auto recovery mode when AXI Bus is congested for RZ/V2H
Below phenomenon is seen for RZ/G2L Series when video bus is heavy and
it is meet with RZ/V2H:
Sometimes, AXI bus congestion caused the image data being corrupted.
This patch is created to set auto recovery mode when AXI Bus is
congested.
It will stop AXI master and clear the FIFO.
After recovery, transfer resumes to the start address of the Memory Bank
to which data was being transferred to.

Because the detection of this issue is recognized by looping frame with
a same HW memory bank, we update minimum buffer to 2 buffers to detect
this issue.

So, apply the same workaround for RZ/V2H.

Signed-off-by: hienhuynh <hien.huynh.px@renesas.com>