arm64: dts: renesas: r9a09g077m44-dev: Modify the macro configuration to enable the settings for PCIe, RSPI and ETHSS
PCIe conflicts with ETH3 pins, and RSPI conflicts with ETH2 pins.
Previously, when using PCIe or RSPI, all ETH channels (ETH0~3)
would be disabled. This patch modifies the macro configuration
settings so that when using PCIe or RSPI, only the ETH channels
with conflicting pins are disabled, while the others remain enabled.

Signed-off-by: Hoa Huynh <hoa.huynh.ud@renesas.com>